data: add Neoverse N2 data
The first time we release a product like N2 r0p0, the filename is just
"neoverse-n2.json". Subsequent revisions use the same file up until a
new core with changes is released. At this point, a file for the new
revision is created along with the files for all the available
revisions.

In this N2 example, up until r0p3 is released, we only have one file:
neoverse-n2.json. When r0p3 is added, we create files for r0p0 and r0p1
as duplicates of the original neovers-n2.json file. A new
neoverse-n2-r0p3.json is created and any subsquent versions of the
product will use that data.

Signed-off-by: Ryan Harkin <ryan.harkin@arm.com>
Change-Id: Ie158d83ab05411d3008f666d3cb1ba6dd003983d