topdown_tool: detect the number of available PMU counters automatically
Automatically detect the number of PMU counters when encountering a
counting error. In a virtual machine environment, the number of
available counters may be less than six.

The detection method on Linux is to start from collecting simultaneously
from 6 events, and decrease one by one, until all events can be completely
collected. And on Windows, directly execute the `wperf test` to retrieve
the number of PMU counters.

It is worth noting that the number of PMU counters can also be obtained
by reading the PMCR_EL0 register, but by default it cannot be read in
userspace(Enabling this feature requires modifying the value of the
PMUSERENR_EL0 register under EL1).

Signed-off-by: Xin Yang <Xin.Yang@arm.com>
Change-Id: I3135bef730e91cb40e64475ede3a80fd08c97800