Adding support to scale generated Perf metrics.
This can be used to apply a scale to metrics to improve precision.
Metrics can be added to tools/perf_json_generator/metric_adjustments.json
Selected metrics have already been added to this file.
The scale_value and scale_units will replace the default scale and units
respectively.
If, only the scale_value is specified, scale_units will be created using
scale_value and default units.
An example of a use case is converting ratio metrics to percentages by using
a scale of 100.

Change-Id: Iea14e256d672161040d3dec12fc1158975efab06