topdown_tool: Re-use test data for similar CPUs
Re-use test data for CPUs where output is expected to be the same.

e.g. tool output is expected to be the same for Neoverse N2 and V2, so
we can test both against the same test data.

This allows removal of duplicate test data, and ensures that we see the
same output for similar CPUs.

Also enables some Neoverse N3 tests that weren't previously enabled.

Change-Id: I8bc07bdf34b720ea434906df74128fb3e9152b9f