perf_json_generator: Fix new common events generation
There was a small mistake in this line:

  present_perf_common_events = [e for e in perf_cpu_events

Instead of returning a subset of 'perf_common_events' filtered on
'perf_cpu_events' it would return a filtered version of
'perf_cpu_events'. This undid the 'describe_common' transformation that
was done at the beginning to use the short title for common events.

Fix it by flipping it to return a subset of common events. This results
in common events having the same title for both BriefDescription and
PublicDescription. Fix that by deleting describe_common() and having
a mrs_to_common_perf_event() instead.

Change-Id: I1d373df92808c5a1ece241a8e30538367f5ceccf