perf_json_generator: Don't duplicate description
The telemetry-solution source has both title and description that can be
used for short and long strings. This is better than repeating the long
one and more closely matches what's already done in Perf. These will
both appear in the CPU's output file if it's not a common event. If it's
a common event, only 'title' gets written into the common file and the
CPU's file gets the event created with only the long 'description' and
an ArchStdEvent reference (see replace_arch_std_event()).

Make the common event reference replacement more robust by handling
either available brief or public descriptions.

Change-Id: Ic220c9ac079147984681666248cd30718e1b19fd