spe_parser: Add late prefetch events
The SPE event packet contains a bitfield indicating late prefetch[1],
which means a memory access hits a cache line whose data is being
prefetched but not ready yet. Such a memory access is counted as a cache
hit, but the wait for data readiness can take many CPU cycles to
complete and potentially stall execution.

One example cause of a high rate of late prefetches is when the
hardware prefetcher makes correct predictions of memory accesses but a
higher level of cache or the DRAM is overloaded and cannot keep up with
the rate of data requests.

This commit adds the missing late prefetch event type.

[1]: https://developer.arm.com/documentation/101427/0102/Debug-descriptions/Statistical-Profiling-Extension/implementation-defined-features-of-SPE

Change-Id: I014b077bfb2ef6d57fd511963b93fc1109764126