TDT-CPU: Add PID-based tracking for homogeneous CPU systems
Previously, topdown-tool captured counters on a per-CPU basis.  
This is necessary for heterogeneous systems where PMU counters  
differ between core types.  

On homogeneous systems with hundreds of cores,this may produces large 
amounts of noise data since all cores participate in the capture and
workloads may likely not run on all of them. 
In this case, it is more practical to capture only the counters on the 
CPU running the workload being analyzed.  

This patch enables PID-based capture when:  
- The system exposes homogeneous CPU cores  
- No specific CPUs were requested by the user  
- A workload or PID was provided to topdown-tool