TDT-WPERF: Heterogeneous core support
Introduce the concept of coordinator to WindowsPerf implementation to
capture data on heterogeneous system.

Wperf do not allow multiple instances to be spawned at once. The
coordinator solves this by serializing request from multiple instances
of WindowsPerf class into a single command. When the capture ends, the
coordinator de-serialize the results back to the perf instances.