TDT-CPU: Extend Linux PMU availability detection
LinuxPerf uses the Linux perf tool to identify how many PMU counters are available on a core. 
This detection is made by running "perf stat" with a set of counter in a single group. 
If the command fails, it means the the system cannot allocate that many hardware PMU.

The well known PMU used was "instructions". Unfortunately it generates an error on some perf
versions. 

To overcome this limitation, we additionally use the PMU "r8" (spelled code of instructions)
if such failure is detected. 

Note that it is not possible to use only "r8" because it will create an error on older 
versions of Linux perf.